Create bbox_qgis
This script will generate new layers for each layer's bounding box. Each new layer will be visible in your Layers panel with the name of the original layer appended with "_bbox".

Remember to adjust the script if you need to specify different save paths or CRS for the new layers.